Separate gameBoard (history) and userGuess (current)
Combination of 2 approaches:
1) Object {letter: , color: } is necessary due to the x-pass nature
of the coloring algorithm; the colors must be saved somewhere
in between passes. Without the Object, this wouldn't be possible.

2) The gameBoard now only updates upon submitGuess, which separates
the purpose of the userGuess (current guess) from the gameBoard
(track history of guesses). For the active row, we render the
userGuess. Inactive rows render the gameBoard.

This dichotomy will be important for upcoming features.