Revert "shared" folder change
For development purposes I thought it would be more convenient
to have a wordleUtils.js, as well as the data arrays, shared by
the client and server. However, in practice, I'm not sure if
the AWS S3 + EC2 will be able to communicate like this. For now,
I'm reverting it so both the client and server have their own copies.