Extract Game Component to fix Routing issue (huge!)
Prior to this commit, I was conditionally rendering Routes in App.jsx
using the isInGame state, which led to a serious bug in which
the user would attempt to join his previous roomId upon starting a new
game. In this refactor, I have extracted the Game component so it
can be used as a Route element. Now, WaitingRoom, etc. are not rendered
conditionally; they are all just separate Routes. tl;dr don't
render Routes conditionally or you'll get bugs.