1/2 Update the database when games are finished. See description.
Overarching philosophy:

1) Update the db in batches. Once a match is over (ex. 10 rounds),
only then do you update the db. This idea came about when I saw that
we could use the prevPoints, prevSolveTime, etc. Maps to our advantage.
Their totals are perfect to update all at once. There may be some
negative effects (ex. all users leave in a half-completed room and
their stats are not updated), but these might be able to be worked
around. Either way, we can save on a lot of db calls in Private Rooms
by doing this. Public will be the same, of course.

2) Public and Private rooms should behave the same. In other words,
the Game properties, such as countSolved, etc., should be incremented
even though it seems trivial for Public games. Right now, this is not
the case; Private rooms receive special treatment. Ideally - and this
is tied to point #1 - the db should essentially look to the Game
properties as an accurate snapshot for updating all its stats.
This adds some much needed generalization to the dbUpdate functions.

There are a multitude of changes in this commit, but they all follow
this philosophy. Going to try and clean up the changes tomorrow.
Currently just missing a bunch of params in the dbUpdate functions.